+++ graphene_django_cud/mutations/batch_delete.py
@@ -9,7 +9,7 @@ from graphql import GraphQLError
 from graphql_relay import to_global_id
 
 from graphene_django_cud.mutations.core import DjangoCudBase
-from graphene_django_cud.util import to_snake_case
+from graphene_django_cud.str_converters import to_snake_case
 
 
 class DjangoBatchDeleteMutationOptions(MutationOptions):
